ESP32: esp-idf v6.0 fixes for thread based SOCs
- implement the ECDSA_sign_msg using psa APIs for SOCs supporting ecdsa
  hardware signing
- fix a few CHIP_ERROR nodiscard
- handle deprecated event enum IP_EVENT_AP_STAIPASSIGNED for v6.0
- fix the implicit conversion from size_t to uint8_t

+++ src/platform/OpenThread/GenericThreadBorderRouterDelegate.cpp
@@ -218,7 +218,7 @@ CHIP_ERROR GenericOpenThreadBorderRouterDelegate::SetPendingDataset(const Thread
 
     otOperationalDatasetTlvs datasetTlvs;
     memcpy(datasetTlvs.mTlvs, pendingDataset.AsByteSpan().data(), pendingDataset.AsByteSpan().size());
-    datasetTlvs.mLength = pendingDataset.AsByteSpan().size();
+    datasetTlvs.mLength = static_cast<uint8_t>(pendingDataset.AsByteSpan().size());
     {
         ScopedThreadLock threadLock;
         VerifyOrReturnError(otDatasetSetPendingTlvs(otInst, &datasetTlvs) == OT_ERROR_NONE, CHIP_ERROR_INTERNAL);
